Replace "libphutil" with "Arcanist"
Inspired by https://we.phorge.it/D26256; you seem to call the sort of
thing here either a "library" when context is clear or an "Arcanist
Library". Do the latter here.

I'm intentionally not replacing the "Phabricator" instances with
"Phorge" in the same file - since both Wikimedia and Translatewiki
unfortunalely still call the project "Phabricator"
I'd rather not be inconsistent with them.

Change-Id: I4694b3b1c45c8dd3c49f6c3b0d6d2b79155f678c