Add an internal translation file for this repo's own strings
This is a bit silly, but https://we.phorge.it/D26572 requires that every
string containing the placeholder `(s)` include a branch on it in the
US English translation, so add that since I eventually want this repo
to comply with upstream validation conventions.

Especially since the translations don't even load unless
https://we.phorge.it/D26580 has landed and you specify a `--locale` arg,
as the default is proto-English.

(This is in `src/management` rather than `src/translations` because I
want that directory to contain only generated code)

Bug: T412651
Change-Id: I57116cefca0b38ad709df62e7e20d304b4e803ca